livelaw_search/requirements.txt
Flask==2.0.3
elasticsearch==8.1.0
mysql-connector==2.2.9
flask_restful==0.3.9
celery==5.2.6
flasgger==0.9.5
python-dotenv==0.20.0